Move the checksum related routine from FileSystem to checksum, where they belong
