+++ src/common/checksums.cpp
@@ -21,6 +21,11 @@
 
 #include <QLoggingCategory>
 #include <qtconcurrentrun.h>
+#include <QCryptographicHash>
+
+#ifdef ZLIB_FOUND
+#include <zlib.h>
+#endif
 
 /** \file checksums.cpp
  *
@@ -79,6 +84,53 @@
 namespace OCC {
 
 Q_LOGGING_CATEGORY(lcChecksums, "sync.checksums", QtInfoMsg)
+
+#define BUFSIZE qint64(500 * 1024) // 500 KiB
+
+static QByteArray calcCryptoHash( const QString& filename, QCryptographicHash::Algorithm algo )
+{
+     QFile file(filename);
+     QByteArray arr;
+     QCryptographicHash crypto( algo );
+
+     if (file.open(QIODevice::ReadOnly)) {
+         if (crypto.addData(&file)) {
+             arr = crypto.result().toHex();
+         }
+     }
+     return arr;
+ }
+
+QByteArray calcMd5(const QString &filename)
+{
+    return calcCryptoHash(filename, QCryptographicHash::Md5);
+}
+
+QByteArray calcSha1(const QString &filename)
+{
+    return calcCryptoHash(filename, QCryptographicHash::Sha1);
+}
+
+#ifdef ZLIB_FOUND
+QByteArray calcAdler32(const QString &filename)
+{
+    QFile file(filename);
+    const qint64 bufSize = qMin(BUFSIZE, file.size() + 1);
+    QByteArray buf(bufSize, Qt::Uninitialized);
+
+    unsigned int adler = adler32(0L, Z_NULL, 0);
+    if (file.open(QIODevice::ReadOnly)) {
+        qint64 size;
+        while (!file.atEnd()) {
+            size = file.read(buf.data(), bufSize);
+            if (size > 0)
+                adler = adler32(adler, (const Bytef *)buf.data(), size);
+        }
+    }
+
+    return QByteArray::number(adler, 16);
+}
+#endif
 
 QByteArray makeChecksumHeader(const QByteArray &checksumType, const QByteArray &checksum)
 {
@@ -188,13 +240,13 @@ QByteArray ComputeChecksum::computeNow(const QString &filePath, const QByteArray
     }
 
     if (checksumType == checkSumMD5C) {
-        return FileSystem::calcMd5(filePath);
+        return calcMd5(filePath);
     } else if (checksumType == checkSumSHA1C) {
-        return FileSystem::calcSha1(filePath);
+        return calcSha1(filePath);
     }
 #ifdef ZLIB_FOUND
     else if (checksumType == checkSumAdlerC) {
-        return FileSystem::calcAdler32(filePath);
+        return calcAdler32(filePath);
     }
 #endif
     // for an unknown checksum or no checksum, we're done right now
