+++ test/CMakeLists.txt
@@ -39,7 +39,6 @@ owncloud_add_test(ChecksumValidator "")
 
 owncloud_add_test(ExcludedFiles "")
 
-owncloud_add_test(FileSystem "")
 owncloud_add_test(Utility "")
 owncloud_add_test(SyncEngine "syncenginetestutils.h")
 owncloud_add_test(SyncVirtualFiles "syncenginetestutils.h")
